Moved register enum to own file
