+++ src/virtual_machine/mod.rs
@@ -1,3 +1,4 @@
 pub mod instructions;
 pub mod lc3_virtual_machine;
+pub mod register;
 pub mod trap;
